+++ examples/js/app.js
@@ -23,6 +23,7 @@ import Complex from './complex/demo';
 import Remote from './remote/demo';
 import Expand from './expandRow/demo';
 import Custom from './custom/demo';
+import Span from './column-header-span/demo';
 
 const renderApp = () => {
   ReactDOM.render(
@@ -37,6 +38,7 @@ const renderApp = () => {
             <Route path='sort' component={ Sort } />
             <Route path='column-format' component={ ColumnFormat } />
             <Route path='column-filter' component={ ColumnFilter } />
+            <Route path='column-header-span' component={ Span } />
             <Route path='selection' component={ Selection } />
             <Route path='pagination' component={ Pagination } />
             <Route path='manipulation' component={ Manipulation } />