+++ examples/js/style/edit-td-class-table.js
@@ -1,7 +1,5 @@
 /* eslint max-len: 0 */
-/* eslint no-alert: 0 */
-/* eslint guard-for-in: 0 */
-/* eslint no-console: 0 */
+/* eslint no-unused-vars: 0 */
 import React from 'react';
 import { BootstrapTable, TableHeaderColumn } from 'react-bootstrap-table';
 
@@ -20,30 +18,18 @@ function addProducts(quantity) {
   }
 }
 
-addProducts(50);
+addProducts(5);
 
-export default class SelectAllOnAllPage extends React.Component {
-
-  onSelectAll = (isSelected) => {
-    if (isSelected) {
-      return products.map(row => row.id);
-    } else {
-      return [];
-    }
-  }
-
+export default class EditColumnClassTable extends React.Component {
   render() {
-    const selectRowProp = {
-      mode: 'checkbox',
-      clickToSelect: true,
-      onSelectAll: this.onSelectAll
+    const cellEditProps = {
+      mode: 'click'
     };
-
     return (
-      <BootstrapTable ref='table' data={ products } selectRow={ selectRowProp } pagination>
+      <BootstrapTable data={ products } cellEdit={ cellEditProps }>
           <TableHeaderColumn dataField='id' isKey={ true }>Product ID</TableHeaderColumn>
           <TableHeaderColumn dataField='name'>Product Name</TableHeaderColumn>
-          <TableHeaderColumn dataField='price'>Product Price</TableHeaderColumn>
+          <TableHeaderColumn dataField='price' editColumnClassName='class-for-editing-cell'>Product Price</TableHeaderColumn>
       </BootstrapTable>
     );
   }