+++ src/TableBody.js
@@ -1,4 +1,5 @@
 import React, { Component, PropTypes } from 'react';
+import Utils from './util';
 import Const from './Const';
 import TableRow from './TableRow';
 import TableColumn from './TableColumn';
@@ -32,9 +33,13 @@ class TableBody extends Component {
     const noneditableRows = (cellEdit.nonEditableRows && cellEdit.nonEditableRows()) || [];
     const unselectable = this.props.selectRow.unselectable || [];
     const isSelectRowDefined = this._isSelectRowDefined();
-    const tableHeader = this.renderTableHeader(isSelectRowDefined);
+    const tableHeader = Utils.renderColGroup(this.props.columns, this.props.selectRow, 'header');
     const inputType = this.props.selectRow.mode === Const.ROW_SELECT_SINGLE ? 'radio' : 'checkbox';
     const CustomComponent = this.props.selectRow.customComponent;
+    let expandColSpan = this.props.columns.filter(col => !col.hidden).length;
+    if (isSelectRowDefined && !this.props.selectRow.hideSelectColumn) {
+      expandColSpan += 1;
+    }
 
     const tableRows = this.props.data.map(function(data, r) {
       const tableColumns = this.props.columns.map(function(column, i) {
@@ -65,7 +70,8 @@ class TableBody extends Component {
                 rowIndex={ r }
                 colIndex={ i }
                 row={ data }
-                fieldValue={ fieldValue } />
+                fieldValue={ fieldValue }
+                className={ column.editClassName } />
             );
         } else {
           // add by bluespring for className customize
@@ -98,7 +104,8 @@ class TableBody extends Component {
               cellEdit={ cellEdit }
               hidden={ column.hidden }
               onEdit={ this.handleEditCell }
-              width={ column.width }>
+              width={ column.width }
+              onClick={ this.handleClickCell }>
               { columnChild }
             </TableColumn>
           );
@@ -129,17 +136,12 @@ class TableBody extends Component {
       </TableRow> ];
 
       if (this.props.expandableRow && this.props.expandableRow(data)) {
-        let colSpan = this.props.columns.length;
-        const bgColor = this.props.expandRowBgColor || this.props.selectRow.bgColor || undefined;
-        if (isSelectRowDefined && !this.props.selectRow.hideSelectColumn) {
-          colSpan += 1;
-        }
         result.push(
           <ExpandComponent
             className={ trClassName }
-            bgColor={ bgColor }
+            bgColor={ this.props.expandRowBgColor || this.props.selectRow.bgColor || undefined }
             hidden={ !(this.state.expanding.indexOf(key) > -1) }
-            colSpan={ colSpan }
+            colSpan={ expandColSpan }
             width={ "100%" }>
             { this.props.expandComponent(data) }
           </ExpandComponent>
@@ -164,45 +166,12 @@ class TableBody extends Component {
         className={ classSet('react-bs-container-body', this.props.bodyContainerClass) }
         style={ this.props.style }>
         <table className={ tableClasses }>
-          { tableHeader }
+          { React.cloneElement(tableHeader, { ref: 'header' }) }
           <tbody ref='tbody'>
             { tableRows }
           </tbody>
         </table>
       </div>
-    );
-  }
-
-  renderTableHeader(isSelectRowDefined) {
-    let selectRowHeader = null;
-
-    if (isSelectRowDefined) {
-      const style = {
-        width: 30,
-        minWidth: 30
-      };
-      if (!this.props.selectRow.hideSelectColumn) {
-        selectRowHeader = (<col style={ style } key={ -1 }></col>);
-      }
-    }
-    const theader = this.props.columns.map(function(column, i) {
-      const style = {
-        display: column.hidden ? 'none' : null
-      };
-      if (column.width) {
-        const width = parseInt(column.width, 10);
-        style.width = width;
-        /** add min-wdth to fix user assign column width
-        not eq offsetWidth in large column table **/
-        style.minWidth = width;
-      }
-      return (<col style={ style } key={ i } className={ column.className }></col>);
-    });
-
-    return (
-      <colgroup ref='header'>
-        { selectRowHeader }{ theader }
-      </colgroup>
     );
   }
 
@@ -217,37 +186,13 @@ class TableBody extends Component {
   }
 
   handleRowClick = rowIndex => {
-    let selectedRow;
-    const { data, onRowClick } = this.props;
-    data.forEach((row, i) => {
-      if (i === rowIndex - 1) {
-        selectedRow = row;
-      }
-    });
-    const rowKey = selectedRow[this.props.keyField];
-    if (this.props.expandableRow) {
-      let expanding = this.state.expanding;
-      if (this.state.expanding.indexOf(rowKey) > -1) {
-        expanding = expanding.filter(k => k !== rowKey);
-      } else {
-        expanding.push(rowKey);
-      }
-      this.setState({ expanding }, () => {
-        this.props.adjustHeaderWidth();
-      });
-    }
-    onRowClick(selectedRow);
+    this.props.onRowClick(this.props.data[rowIndex - 1]);
   }
 
   handleRowDoubleClick = rowIndex => {
-    let selectedRow;
-    const { data, onRowDoubleClick } = this.props;
-    data.forEach((row, i) => {
-      if (i === rowIndex - 1) {
-        selectedRow = row;
-      }
-    });
-    onRowDoubleClick(selectedRow);
+    const { onRowDoubleClick } = this.props;
+    const targetRow = this.props.data[rowIndex];
+    onRowDoubleClick(targetRow);
   }
 
   handleSelectRow = (rowIndex, isSelected, e) => {
@@ -269,6 +214,35 @@ class TableBody extends Component {
         rowIndex + 1,
         e.currentTarget.checked,
         e);
+    }
+  }
+
+  handleClickCell = (rowIndex, columnIndex) => {
+    const {
+      columns,
+      keyField,
+      expandBy,
+      expandableRow,
+      selectRow: {
+        clickToExpand
+      }
+    } = this.props;
+    const selectRowAndExpand = this._isSelectRowDefined() && !clickToExpand ? false : true;
+
+    if (expandableRow &&
+      selectRowAndExpand &&
+      (expandBy === Const.EXPAND_BY_ROW ||
+      (expandBy === Const.EXPAND_BY_COL && columns[columnIndex].expandable))) {
+      const rowKey = this.props.data[rowIndex - 1][keyField];
+      let expanding = this.state.expanding;
+      if (expanding.indexOf(rowKey) > -1) {
+        expanding = expanding.filter(k => k !== rowKey);
+      } else {
+        expanding.push(rowKey);
+      }
+      this.setState({ expanding }, () => {
+        this.props.adjustHeaderWidth();
+      });
     }
   }
 
@@ -339,6 +313,7 @@ TableBody.propTypes = {
   expandableRow: PropTypes.func,
   expandComponent: PropTypes.func,
   expandRowBgColor: PropTypes.string,
+  expandBy: PropTypes.string,
   adjustHeaderWidth: PropTypes.func
 };
 export default TableBody;