+++ src/csv_export_util.js
@@ -13,7 +13,34 @@ function toString(data, keys) {
   let dataString = '';
   if (data.length === 0) return dataString;
 
-  dataString += keys.map(x => x.header).join(',') + '\n';
+  const headCells = [];
+  let rowCount = 0;
+  keys.forEach(key => {
+    if (key.row > rowCount) {
+      rowCount = key.row;
+    }
+    // rowCount += (key.rowSpan + key.colSpan - 1);
+    for (var index = 0; index < key.colSpan; index++) {
+      headCells.push(key);
+    }
+  });
+
+  for (let i = 0; i <= rowCount; i++) {
+    dataString += headCells.map(x => {
+      if ((x.row + (x.rowSpan - 1)) === i) {
+        return x.header;
+      }
+      if (x.row === i && x.rowSpan > 1) {
+        return '';
+      }
+    }).filter(key => {
+      return typeof key !== 'undefined';
+    }).join(',') + '\n';
+  }
+
+  keys = keys.filter(key => {
+    return key.field !== undefined;
+  });
 
   data.map(function(row) {
     keys.map(function(col, i) {